+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Added
 
+- emails: add `get_from_email()` hook on `EmailBase` so projects can customize
+  the sender address without overriding `dispatch()`.
+
 ### Fixed
 
 ### Changed